+++ Modules/Temperature/LM35.h
@@ -5,7 +5,7 @@
 #include "../DIO/DIO.h"
 
 void LM35_init(void);
-uint32_t GetTemp(void);
+double GetTemp(void);
 void tempEn_init(char port, uint8_t pins);
 void Read_Temperature_Handler(void);
 void tempEN(char c);
